renames around legacy-default-xpring-client
